% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alpha.R
\name{determine_alpha_fit}
\alias{determine_alpha_fit}
\title{Calculate coefficients needed to compute alpha and spread}
\usage{
determine_alpha_fit(scalelength = 5, samples = 5, data = NULL, ...)
}
\arguments{
\item{scalelength}{How many items are in the simulated scale? This should match the Likert Scale you are trying to replicate.}

\item{samples}{How many runs should be undertaken? More is generally better and will result in less noise at low consistency, but this is at the expense of speed.}

\item{data}{optional dataframe of class alphatable to speed up processing by reusing pre-generated data}
}
\value{
an object of class alphafit
}
\description{
This helper function simulates cases and determines the parameters of the best fit curve needed to compute alpha. This function looks for a data frame generated by make_alpha_table and if it can't be found generates this.
}
\examples{

determine_alpha_fit()

}
