% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alpha.R
\name{estimate_alpha}
\alias{estimate_alpha}
\title{Estimate the implied value of alpha from rho}
\usage{
estimate_alpha(rho, model)
}
\arguments{
\item{rho}{The desired value of rho to test}

\item{model}{A model created by \code{determine_alpha_fit()}}
}
\value{
A numeric value for alpha
}
\description{
In \code{generateDF()} and \code{generateData()} the parameter \code{rho} controls the variability within the responses for \strong{each} respondent(case) if a multi item scale is being created. Rho corresponds to the standard deviation of the distribution used to generate the individual item responses and is dependent on the number of items in the Likert scale and the intended internal consistency (measured using Cronbach alpha for simplicity). Sometimes it is necessary to back calculate the value of alpha for a model from a chosen rho. This function allows for a value of alpha to be determined relatively easily.
}
\examples{
m <- determine_alpha_fit(data = x)
estimate_alpha(1.46, m)

}
\seealso{
\code{\link[=estimate_rho]{estimate_rho()}}

\code{\link[=determine_alpha_fit]{determine_alpha_fit()}}

\code{\link[=generateData]{generateData()}}

\code{\link[=plot_rho_curve]{plot_rho_curve()}}
}
