% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alpha.R
\name{estimate_rho}
\alias{estimate_rho}
\title{Estimate the necessary value of rho}
\usage{
estimate_rho(alpha, model)
}
\arguments{
\item{alpha}{The desired level of Cronbach alpha to simulate}

\item{model}{A model created by \code{determine_alpha_fit()}}
}
\value{
A numeric value for rho
}
\description{
In \code{generateDF()} and \code{generateData()} the parameter \code{rho} controls the variability within the responses for \strong{each} respondent(case) if a multi item scale is being created. This parameter corresponds to the standard deviation of the distribution used to generate the individual item responses and is dependent on the number of items in the Likert scale and the intended internal consistency (measured using Cronbach alpha for simplicity). This function allows for a value of rho to be determined relatively easily.

\strong{NOTE:} The underlying model generation function \code{determine_alpha_fit()} is quite noisy for alpha < 0.4. Increasing \code{samples} in that function will increase precision at the expense of speed.
}
\details{
While Cronbach alpha is often considered to be an OK measure of internal consistency, it is often abused as a measure of reliability (see details section of \code{\link[psych:alpha]{psych::alpha()}} for a good discussion on the topic). We use it here purely for convenience and due its familiarity with users. Various cut off points have been suggested, and these do vary quite widely with field.  We suggest the following cut offs and qualitative descriptions as a rule of thumb when simulating data sets\tabular{cl}{
   Cronbach Alpha \tab Interpretation \cr
   ≥0.9 \tab Excellent consistency \cr
   ≥0.8 \tab Good level of consistency \cr
   ≥0.7 \tab Reasonable level of consistency \cr
   ≥0.6 \tab Questionable level of consistency \cr
   <0.6 \tab Concerning level of consistency or Inconsistent \cr
}
}
\examples{
m <- determine_alpha_fit()
estimate_rho(alpha = 0.8, model = m)

}
\seealso{
\code{\link[=determine_alpha_fit]{determine_alpha_fit()}}

\code{\link[=generateData]{generateData()}}

\code{\link[=plot_rho_curve]{plot_rho_curve()}}

\code{\link[=estimate_alpha]{estimate_alpha()}}
}
