% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate Likert Data}
\usage{
generateData(
  n = 10,
  itemlength = 5,
  scalewidth = 5,
  mu = NULL,
  delta = NULL,
  rho = 1.46,
  sigma = 2,
  trim = TRUE,
  identifiers = c("Control", "Intervention"),
  effect = NULL,
  seed = NULL,
  type = "mean",
  return = "summary"
)
}
\arguments{
\item{n}{The number of cases to generate for each group}

\item{itemlength}{How many choices should the Likert item have? Defaults to 5.}

\item{scalewidth}{How many items should be generated to represent the Likert scale? Defaults to 5}

\item{mu}{The intended mean response of the first group. If not set, then the centre of the item length is used.}

\item{delta}{If simulation of misunderstanding the scale choice points is desired, then this sets the standard deviation of the uncertainty in the data.}

\item{rho}{The standard deviation of the item responses \strong{per} case. This corresponds to variability \strong{within} each respondent's scores}

\item{sigma}{How many standard deviations either side of mu should the width of the Likert item choices really represent?}

\item{trim}{What should be done with out of range values. If TRUE then out of range values will be replaced with NA. If FALSE, then out of range values will be trimmed to the min of max item choices as appropriate.}

\item{identifiers}{A vector of strings corresponding to group identifiers. The first three letters will be manipulated to create case identifiers. The first value will always be used. subsequent values will only be used if effect is not NULL}

\item{effect}{A numeric vector corresponding to the intended effect sizes for each of the groups other than the first group. This should have length 1 less than identifiers.}

\item{seed}{optional seed for the random number generator}

\item{type}{Either "mean" or "sum" that describes the type of summary data returned. Note that if set to sum then trim will automatically be set to FALSE}

\item{return}{One of "all", "raw", "summary", "uncertainty" (first letter will work). This determines the nature of the returned dataframe.}
}
\value{
A data frame with structure controlled by the value of return
}
\description{
Generates a data set of simulated Likert data based on well-defined parameters. The centrality, and distribution of responses can be set for the generated data as well as the number of items in the scale and the number of response points for the Likert items. It is possible to also generate data sets for control and one or more intervention groups defined by an effect size, and to also generate data that includes a degree of additional uncertainty representing respondents misinterpretting the meaning of scale points.
}
\examples{
##Create a simple data frame of 5 cases
generateData(n=5)

##Create a simple data frame of 5 cases for each of boys and girls where the 
##boys' responses are 1sd greater than the girls'.
generateData(n=5,
             effect=1,
             identifiers=c("girls","boys"))

##Create a simple dataframe with summed outputs 
generateData(n = 5, 
             scalewidth = 3, 
             effect = 1, 
             identifiers = c("bees","birds"), 
             type = "sum")

##Create a dataframe of 5 cases for each of humans and aliens where the aliens' 
##responses are 1 sd lower than the humans and fill in any NAs with min and max 
##values. Generate items corresponding to a 4 point Likert item and have 7 items 
##in the scale
generateData(n=5, 
             itemlength = 4, 
             scalewidth = 7, 
             effect = -1, 
             identifiers=c("humans","aliens"), 
             trim = FALSE, 
             return = "raw")

}
