% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDataFrame.R
\name{generateDF}
\alias{generateDF}
\title{GenerateDF}
\usage{
generateDF(
  n = 10,
  itemlength = 5,
  scalewidth = 5,
  mu = NULL,
  delta = NULL,
  rho = 1.46,
  sigma = 2,
  identifier = "",
  seed = NULL,
  trim = TRUE
)
}
\arguments{
\item{n}{The number of rows to generate}

\item{itemlength}{How many scale points does the Likert scale have?}

\item{scalewidth}{How many items make up the full scale}

\item{mu}{The intended mean scale point of the returned data set. If not set, then the centre of the scale is used.}

\item{delta}{If uncertainty related to misunderstanding the scale items is desired, then this sets the magnitude of the standard deviation of that uncertainty}

\item{rho}{The standard deviation of the item responses \strong{per} row. This corresponds to variability within the items on a per respondent level.}

\item{sigma}{how many standard deviations should the full range really represent}

\item{identifier}{An optional string that is prepended to id variables}

\item{seed}{optional seed for the random number generator}

\item{trim}{If TRUE then out of range values will be replaced by NA. If FALSE then out of range values will be set to the min and max values as appropriate}
}
\value{
A wide format dataframe of simulated response values (x) and values with additional uncertainty (e) if delta is set.
}
\description{
Generates a dataframe of variable width that corresponds to \strong{raw} Likert data
}
\examples{
##Create a simple dataframe for a 5-item Likert scale with each item on a scale from 1 to 5
generateDF(identifier = "C")

##As above but replace NA with min and max scale values
generateDF(trim = FALSE, 
           identifier = "C")

##As example 1, but simulate 32\% of participants misinterpreting scale points
generateDF(delta = 1, 
           identifier = "C")

##Create a 6-item scale, of 7-point Likert items centred on a response of 5 and no uncertainty.
generateDF(n=20, 
           itemlength = 7, 
           scalewidth = 6, 
           mu =5, 
           identifier = "Student")

}
