% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alpha.R
\name{is.alphafit}
\alias{is.alphafit}
\title{Checking data classes of alphafit}
\usage{
is.alphafit(x)
}
\arguments{
\item{x}{any R object}
}
\value{
returns TRUE it its argument is an alphafit model or FALSE otehrwise
}
\description{
Check if an object contains all the necessary information to be an alphafit model
}
\examples{
x <- tibble::tibble(n = seq(1,100))
is.alphafit(x)

x <- determine_alpha_fit()
is.alphafit(x)

}
