% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_variables.R
\name{label_variables}
\alias{label_variables}
\title{Label Variables with Human Readable Names}
\usage{
label_variables(.data, scale = "agreement", labels = NULL)
}
\arguments{
\item{.data}{A data frame containing columns named x## e## or Response created by generateData()}

\item{scale}{One of \strong{AG}reement, \strong{AP}proval, or \strong{TE}mporal}

\item{labels}{An optional vector of labels to apply to the factor levels.}
}
\value{
A data frame of the same form as the input with numbers replaced by labels
}
\description{
Label Variables with Human Readable Names
}
\examples{
generateData(return = "raw")\%>\%
  label_variables()
  
generateData(return = "all", delta = 0.2, scalewidth = 3)\%>\%
  label_variables(scale = "TE") 
 
generateData(return = "all", scalewidth = 3, itemlength = 7)\%>\%
  label_variables(scale = "AP")   
  
generateData(return = 'raw')\%>\%
  label_variables(labels = c("a","b","c","d","e"))
}
