% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alpha.R
\name{make_alpha_table}
\alias{make_alpha_table}
\title{Compute table of data spread and Cronbach alpha}
\usage{
make_alpha_table(scalelength = 5, samples = 5, cases = 5000)
}
\arguments{
\item{scalelength}{How many items are in the simulated scale? This should match the Likert Scale you are trying to replicate.}

\item{samples}{How many runs should be undertaken? More is generally better and will result in less noise at low consistency, but this is at the expense of speed.}

\item{cases}{How many cases should be generated for each simulation run.}
}
\value{
a 2 column tibble of class alphatable
}
\description{
Alpha, a commonly accepted measure of internal consistency needs to be computed from actual data rather than theory. Data that have very little variability within a case have high internal consistency. This helper function creates a series of simulated data points with random noise added drawn from a normal distribution of varying standard distribution from 0 to 4. Alpha is then calculated for each set of results and a summary table is returned.  This is a slow function and it is very unlikely you actually want to be calling it.
}
\examples{
make_alpha_table()

make_alpha_table(scalelength = 7, samples = 8)

}
