% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alpha.R
\name{plot_alpha_curve}
\alias{plot_alpha_curve}
\title{Draw a plot showing the quality of the fit between}
\usage{
plot_alpha_curve(x, ...)
}
\arguments{
\item{x}{a dataset of class alphatable}

\item{...}{values passed to other functions}
}
\value{
ggplot2 object
}
\description{
Draw a plot showing the quality of the fit between
}
\examples{
\donttest{
x <- make_alpha_table()
plot_alpha_curve(x)
}

}
