% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alpha.R
\name{plot_rho_curve}
\alias{plot_rho_curve}
\title{Plot the Alpha to Rho Calibration curve}
\usage{
plot_rho_curve(model)
}
\arguments{
\item{model}{an alphafit model}
}
\value{
a ggplot2 object
}
\description{
Rho for use in \code{generateData()} is determined by simulation and the generation of an approximate equation essentially making use of a calibration curve. Use this function to show that calibration curve
}
\examples{
\donttest{
m <- determine_alpha_fit()
plot_rho_curve(m)
}

}
