% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alpha.R
\name{print.alphafit}
\alias{print.alphafit}
\title{Generic Print Method for class(alphafit)}
\usage{
\method{print}{alphafit}(x, ...)
}
\arguments{
\item{x}{an object of class alphafit}

\item{...}{further arguments passed to or from other methods}
}
\value{
Output to StdOut
}
\description{
Generic Print Method for class(alphafit)
}
\examples{

x <- determine_alpha_fit()
print(x)
}
