% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_numeric.R
\name{to_numeric}
\alias{to_numeric}
\title{to_numeric}
\usage{
to_numeric(.data, match = "^(x|e)[1-9]+$|^Response$")
}
\arguments{
\item{.data}{A data frame generated by \code{generateData()}}

\item{match}{A RegEx to match. By default \verb{^(x|e)[1-9+$]}}
}
\value{
A matching data frame
}
\description{
Converts Likert data generated by \code{generateData()} from factor form to its underlying numeric format.  By default works on any columns that match the RegEx in \code{match}.
#'
}
\examples{
generateData(return = "raw")\%>\%
  to_numeric()
  
generateData(return = "raw")\%>\%
  label_variables()\%>\%
  to_numeric()

}
