% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise}
\alias{visualise}
\alias{visualize}
\title{Visualise Likert Data}
\usage{
visualise(.data, vars = NULL, discrete = FALSE)

visualize(.data, vars = NULL, discrete = FALSE)
}
\arguments{
\item{.data}{A data frame generated by \code{generateData()}. Can be in wide or long format}

\item{vars}{An optional character vector or list of variables to include in the display}

\item{discrete}{An optional boolean to try to force the output to be a bar chart and not a density plot. Be careful when using as results may be unexpected.}
}
\value{
A ggplot2 object
}
\description{
A quick wrapper to visualize data generated by \code{GenerateData()} in a way that allows adjustments to be made easily. If data are numeric then density plots are returned, if data are factors then bar plot is returned.
}
\examples{

# Generate 1000 cases and show the average response
generateData(n=1000)\%>\%
  visualise()

# Generate 100 cases for an intervention of effect size 1 and only show data from x1 and x2 
generateData(n=100, return = "raw", effect = 1)\%>\%
  visualise(vars = c("x1","x2"))
  
# Generate 500 cases for an intervention of effect size -1 and only show data from x1 and 
# applying Temporal Lables 
generateData(n=100, return = "raw", effect = 1)\%>\%
  label_variables(scale = "TE")\%>\%
  visualise(vars = c("x1"))
}
