% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_longer.R
\name{make_longer}
\alias{make_longer}
\title{make_longer()}
\usage{
make_longer(.data)
}
\arguments{
\item{.data}{A wide dataframe generated by generateData()}
}
\value{
A long dataframe
}
\description{
Make a dataframe generated by \code{generateData()} into long form following tidy rules
}
\examples{
generateData(n = 5, 
             scalewidth = 3, 
             return = "raw")

generateData(n = 5, 
            scalewidth = 3, 
            return = "raw")\%>\%
 make_longer()
}
